pyproject.toml
[tool.poetry]
name = "social-events-api"
version = "0.1.0"
description = "Create and manage events for group of friends"
authors = ["Karol Gajda <karol.gajda97@gmail.com>"]
readme = "README.md"
packages = [{include = "social_events_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^3.8.1"
celery = "^5.2.7"
drf-spectacular = "^0.25.1"
dj-database-url = "^1.2.0"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
django-recurrence = "^1.11.1"
django-light = "^0.1.0.post3"
django-filter = "^22.1"
dj-rest-auth = "^3.0.0"
django-allauth = "^0.52.0"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
mypy = "^0.991"
django-stubs = "^1.13.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"